% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{add_region_ID}
\alias{add_region_ID}
\title{Match GCAM ID to region using data from a lookup table.}
\usage{
add_region_ID(datatable, lookupfile = lut.rgn32, provincefile = NULL,
  drops = NULL)
}
\arguments{
\item{datatable}{A table of results produced by \code{\link[rgcam]{getQuery}}}

\item{lookupfile}{Name of one of the predefined map sets, OR, if you're using
a custom map set, the file containing the region lookup table}

\item{provincefile}{Name of one of the predefined map sets, OR, if you're
using a custom map set, file containing the province lookup table, if
applicable.}

\item{drops}{Name of one of the predefined map sets, OR, if you're using
a custom map set, the file containing a list of regions to drop, if
applicable.}
}
\value{
Input table modified to include a GCAM ID for reach region.
}
\description{
We match by ID number to avoid problems with variant spellings and the like.
With the optional arguments you can also omit regions for which you don't
want to plot the data for some reason, and you can translate the
abbreviations used in subregion output.
}
\details{
The \code{provincefile} and \code{drops} arguments are a little clunky.  They
are optional, but if you are using one of the built-in map sets, then you
\emph{must not} specify them if they don't exist for the map set you are
using.  Currently, \code{rgn14} and \code{basin235} have neither drops nor
province abbreviations.  The \code{rgn32} set has drops, but not province
abbreviations.  Only the \code{chn} set (and the \code{usa} set, when it is
finally implemented) has both.
}
