% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{addRegionID}
\alias{addRegionID}
\title{Match GCAM ID to region using data from a lookup table.}
\usage{
addRegionID(datatable, lookupfile, provincefile = "none", drops = "none")
}
\arguments{
\item{datatable}{A table of results produced by \code{\link{process_batch_q}}}

\item{lookupfile}{File containing the region lookup table.  XXX: we
need to provide a mechanism for users to use the data installed
internally in the package.}

\item{provincefile}{File containing the province lookup table, if
applicable. XXX: Same comment as above}

\item{drops}{File containing a list of regions to drop, if
applicable.  XXX: Same comment as above.}
}
\value{
Input table modified to include a GCAM ID for reach region.
}
\description{
We match by ID number to avoid problems with variant spellings and the like.
}

