% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_header.R
\docType{data}
\name{af_ortho}
\alias{af_ortho}
\title{Projection string for orthographic projection over Africa}
\format{An object of class \code{character} of length 1.}
\usage{
af_ortho
}
\description{
String for specifying the orthographic projection over Africa.  You can pass
this value to the \code{proj} argument of \code{\link{plot_GCAM}} to get the
best result.
}
\keyword{datasets}
