% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{assign_prj4s}
\alias{assign_prj4s}
\title{Helper function for to assign return Proj4 string.}
\usage{
assign_prj4s(proj_type, proj)
}
\arguments{
\item{proj_type}{Either esri, epsg, or sr-org as string.  These correspond to
available reference types hosted by http://spatialreference.org/}

\item{proj}{The coordinate reference system number or object provided by the user}
}
\description{
Uses user-input for projection type to identify what
type of URL fetch needs to be conducted to retrieve
a Proj4 string from http://spatialreference.org
}
