% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\docType{data}
\name{basin235}
\alias{basin235}
\title{Designator for the basin235 map set}
\format{An object of class \code{name} of length 1.}
\usage{
basin235
}
\description{
This symbol will select the basin235 map set
}
\keyword{datasets}
