% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{calc.breaks}
\alias{calc.breaks}
\title{Calculate legend breaks (intervals to put labels on legend scale)}
\usage{
calc.breaks(maxval, minval = 0, nbreak = 5, nsig = 3)
}
\arguments{
\item{maxval}{Largest value in the scale}

\item{minval}{Smallest value in the scale}

\item{nbreak}{Number of break points}

\item{nsig}{Number of significant digits to display in the legend.}
}
\description{
Given a minimum and maximum value, and number of breaks, calculate
evenly-spaced break values.
}

