% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\docType{data}
\name{chn}
\alias{chn}
\title{Designator for the chn map set}
\format{An object of class \code{name} of length 1.}
\usage{
chn
}
\description{
This symbol will select the chn map set
}
\keyword{datasets}
