% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{create_map}
\alias{create_map}
\title{Create a map object and return/save the output}
\usage{
create_map(shape_path = NULL, shape_obj = NULL, raster_path = NULL,
  raster_obj = NULL, raster_col = NULL, dpi = 150,
  output_file = NULL, data_classification = "Temp")
}
\arguments{
\item{shape_path}{Input full path string to shape}

\item{shape_obj}{Pass shape object in directly instead of path}

\item{raster_path}{Input full path string to raster}

\item{raster_obj}{Pass raster object in directly instead of path}

\item{dpi}{Settable DPI for different print/screen formats}

\item{output_file}{Output file path to save the resulting plot}

\item{data_classification}{This will be used to determine the colors/units/legend items based on the type of data (ie. temperature, precip, etc)}
}
\value{
A ggplot object of the resulting map
}
\description{
This function is designed to take both a shape and raster object or path and create a standardized, congruent, output map.
}
