% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{custom_map}
\alias{custom_map}
\title{Create a basic map object from input shape/raster and return, save (optional) the output}
\usage{
custom_map(shape_data = NULL, shape_label_field = NULL,
  shape_label_size_field = "1", simplify = FALSE, raster_data = NULL,
  raster_col = NULL, raster_band = 1, convert_zero = FALSE,
  dpi = 150, output_file = NULL, expand_xy = c(0, 0),
  map_xy_min_max = c(-180, 180, -90, 90), map_title = NULL,
  map_palette = "RdYlBu", map_palette_reverse = FALSE,
  map_width_height_in = c(15, 10), map_legend_title = NULL,
  map_x_label = "Lon", map_y_label = "Lat")
}
\arguments{
\item{shape_data}{(SF, SP, or Character) - Either the full path string to a shape file (with included necessary files) or an SF shape object}

\item{shape_label_field}{(Character) - Optional field for plotting data available from the shape attributes/fields (such as country name)}

\item{shape_label_size_field}{(Character) - Optional field used for computing shape label size dynamically (ie by area or amount etc.)}

\item{simplify}{(Boolean) - Option to reduce the number/complexity of the polygons in the shape file (default FALSE)}

\item{raster_data}{(Raster or Character) - Either the full path string to a raster file or an object of type RasterLayer}

\item{raster_col}{(Character) - Column name that contains the raster object's output variable/value}

\item{raster_band}{(Numeric) - Future variable for dealing with multi band/time series rasters etc}

\item{convert_zero}{(Boolean) - Convert values within the raster data from zero to NA (default FALSE)}

\item{dpi}{(Numeric) - Settable DPI for different print/screen formats (default 150)}

\item{output_file}{(Character) - Output file path and file name/type to save the resulting plot (e.g. "c:/temp/output.png") (Types accepted: "eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp", "svg")}

\item{expand_xy}{(c(Numeric, Numeric)) - Sets expansion amount for the X and Y scale of the map - Vector (expand x, expand y) (default c(0,0))}

\item{map_xy_min_max}{(c(Numeric, ...)) - Vector that describes the desired extent of the map in the form of (xmin, xmax, ymin, ymax) (default: c(-180, 180, -90, 90))}

\item{map_title}{(Character) - Title to be displayed on the output map}

\item{map_palette}{(Character) - Variable to hold the type of colorscale to be from the RColorBrewer palette (default "RdYlBu")}

\item{map_palette_reverse}{(Boolean) - Set palette to reverse direction TRUE/FALSE}

\item{map_width_height_in}{(c(Numeric, Numeric)) - Vector that describes the desired file size of the output image in the form of (width, height) in inches (defalt c(15, 10))}

\item{map_legend_title}{(Character) - Text for the legend header}

\item{map_x_label}{(Character) - Label for x axis (default "Lon")}

\item{map_y_label}{(Character) - Label for y axis (default "Lat")}
}
\value{
(ggplot2 or Character) - Returns a ggplot object of the resulting map or an error string if failed
}
\description{
This function is designed to take both a shape and raster object/path and create a standardized output map.
}
