% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam_data_processing.R
\name{drop_regions}
\alias{drop_regions}
\title{Drop regions listed in drops file from data frame.}
\usage{
drop_regions(datatable, drops)
}
\arguments{
\item{datatable}{A data frame containing the output of a GCAM query.}

\item{drops}{String; path to file containing regions to be dropped}
}
\value{
An updated data frame with regions dropped.
}
\description{
Drop regions listed in drops file from data frame.
}
