% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_header.R
\docType{data}
\name{eck3}
\alias{eck3}
\title{Proj4 string for the Eckert III World projection}
\format{An object of class \code{character} of length 1.}
\usage{
eck3
}
\description{
String for specifying the Eckert III projection in mapping
functions.  Its value is \code{'+proj=eck3'}
}
\keyword{datasets}
