% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_header.R
\docType{data}
\name{EXTENT_AFRICA}
\alias{EXTENT_AFRICA}
\title{Extent vector for Africa}
\format{An object of class \code{numeric} of length 4.}
\usage{
EXTENT_AFRICA
}
\description{
This vector can be used as the \code{extent} argument to
\code{\link{plot_GCAM}}.
}
\keyword{datasets}

