% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{filter_spatial}
\alias{filter_spatial}
\title{Get features topologically associated with extent bounds.}
\usage{
filter_spatial(mapdata, bbox, extent, col, topo = sf::st_intersects)
}
\arguments{
\item{mapdata}{The sf object containing the spatial data.}

\item{bbox}{Bounding box.}

\item{extent}{Extent provided by the user or as default.}

\item{col}{Field name with target information to map.}

\item{topo}{SF topologic function to define how the join will be conducted. Default is
to join any feature that intersects the bounding box.}
}
\description{
Conducts a spatial join to retrieve spatial features that are topologically associated
(intersects, contains, within, etc.) with the provided bounds.
}
