% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_header.R
\docType{data}
\name{gcam14_colors}
\alias{gcam14_colors}
\title{Color palette for 14-region GCAM}
\format{An object of class \code{character} of length 14.}
\usage{
gcam14_colors
}
\description{
This palette should be used for plots by region (whether maps, line plots, or
other types) to ensure consistency across plots and publications.  XXX:
Perhaps this sort of thing should go in a separate GCAM style package, since
it isn't really specific to mapping?
}
\keyword{datasets}

