% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_header.R
\docType{data}
\name{gcam32_colors}
\alias{gcam32_colors}
\title{Color palette for 32-region GCAM}
\format{An object of class \code{character} of length 15.}
\usage{
gcam32_colors
}
\description{
This palette should be used for plots by region (whether maps, line plots, or
other types) to ensure consistency across plots and publications.  XXX: This
color palette doesn't seem to be finished yet, so we haven't exported it.
}
\keyword{datasets}

