% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcammaptools.R
\docType{package}
\name{gcammaptools}
\alias{gcammaptools}
\alias{gcammaptools-package}
\title{gcammaptools:  A package for plotting GCAM data on maps}
\description{
The gcammaptools package provides functions for plotting GCAM data on world
or regional maps.  This includes functions for making plots for regional or
gridded data, as well as default projection and theme settings that provide a
house style for GCAM plots.
}
\section{Preparing GCAM data}{


The recommended way to load your GCAM data is by using the \code{rgcam}
package create a project data file from a GCAM database, and then querying
that file for the data you want to plot. Alternatively, you can start with
any data frame that has a `region` column and one or more data columns.

Once you have loaded the data, you must add the region identifiers used in
the map data to the data frame using the \code{add_region_ID} function.
}

\section{Mapping GCAM data}{


To map GCAM data, you will need a simple feature collection or geojson with
your region boundaries.  The package provides the following commonly-used
base maps:
\itemize{
  \item \code{\link{map.rgn14}}: 14 region map used prior to GCAM 4.0.
  \item \code{\link{map.rgn32}}: 32 region map used in GCAM 4.0 and later,
excluding Taiwan.
  \item \code{\link{map.basin235}}: 235 water basins.
  \item \code{\link{map.chn}}: 32 global regions plus China subregions
  \item \code{\link{map.usa}}: 32 global regions plus US states.
}
Users can provide their own base maps.  You will need a geometry file in
geojson format, where each region has a \code{region_id}.  Land masses
that you want to draw, but which aren't part of any region should be included
and assigned a \code{region_id} of 0.  The \code{examples} vignette shows how
to load such a file and convert it for use in this package.

Once you have your map data and your GCAM data, can generate the maps by
passing both to the \code{\link{plot_GCAM}} function.
}

