% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{get.internal}
\alias{get.internal}
\title{Get auxiliary data for a named mapset.}
\usage{
get.internal(mapset, type)
}
\arguments{
\item{mapset}{The name of the mapset.  Can be either a symbol or a string.}

\item{type}{The type of table.  Right now this is either 'lut', 'drop', or
'prov'}
}
\description{
We have several standard map sets.  Each of them has several auxiliary tables
associated with it.  This function retrieves the auxiliary table associated
with the requested.  Right now this function understands \code{rgn14},
\code{rgn32}, \code{basin235}, and \code{chn}.
}
