% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{get_prj4s}
\alias{get_prj4s}
\title{Retrieve proj4 projection string.}
\usage{
get_prj4s(obj = NULL, prj_type = NULL, prj_code = NULL,
  prj4s_key = NULL)
}
\arguments{
\item{obj}{Use object instead that has a predefined proj4 string.}

\item{prj_type}{The projection type is either 'esri', 'epsg', or 'sr-org' or 'prj4s_key'.}

\item{prj_code}{The projection code as an integer for EPSG, ESRI, or SR-ORG.}

\item{prj4s_key}{Lookup key string identifying the default projection.}
}
\description{
Provides a lookup list for default proj4 strings utilized.  Users may also
specify their own lookup list.  Options also include providing either the
EPSG, ESRI, or SR-ORG projection codes to retrieve the associated proj4
string from a web query from http://spatialreference.org.  Definitions for
proj4 string parmeters can be referenced here:
http://proj4.org/parameters.html#parameter-list
}
