% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{import_mapdata}
\alias{import_mapdata}
\title{Single import function for compatible data types.}
\usage{
import_mapdata(obj, fld = NULL,
  prj4s = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs")
}
\arguments{
\item{obj}{Input full path string or object}

\item{fld}{Field name to use as identifier}

\item{prj4s}{Proj4 string for projection (default WGS84)}
}
\description{
Imports available for sf objects, data frames, spatial data frames,
ESRI Shapefiles, GeoJSON files, or text files.
}
