% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{join_gcam}
\alias{join_gcam}
\title{Join GCAM data with spatial data.}
\usage{
join_gcam(mapdata, mapdata_key, gcam_df, gcam_key)
}
\arguments{
\item{mapdata}{The sf object containing the spatial data and a tuple identifier that
can be referenced in the gcam_df data frame.}

\item{mapdata_key}{Name of the field having a tuple identifier that can be referenced
in the gcam_df data frame.}

\item{gcam_df}{The GCAM data frame provided from the user.  This is usually generated from
an rgcam query.}

\item{gcam_key}{Name of field having a tuple identifier that can be referenced in the
mapdata data frame.}
}
\description{
Joins GCAM data from rgam query and inner joins it to spatial data provided by the user.
Note:  due to conducting an inner join, only the keys that are present in both datasets
will be represented.
}
