% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{load_shp}
\alias{load_shp}
\title{Import ESRI Shapefile or GeoJSON as sf object.}
\usage{
load_shp(file_pth = NULL)
}
\arguments{
\item{file_pth}{(Character) Full path to shapefile (.shp). Shapefiles must contain at least .shp, .shx, and .dbf file to function properly.}
}
\value{
(sf or Character) - Returns the loaded SF object or an error string if failed
}
\description{
Creates a Simple Feature (sf) object from full path string to ESRI Shapefile
or GeoJSON file. User defines which field is supposed to represent the ID for
the data.
}
