% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcammaptools.R
\docType{data}
\name{map.basin235}
\alias{map.basin235}
\title{Base map for 235 global water basins}
\format{Simple feature collection}
\usage{
map.basin235
}
\description{
This is the map of the 235 global water basins.  For compatibility with the
other map data frames it refers to the basins as 'regions'.  Thus, any GCAM
data with a 'basin' column will need to have a 'region' column added.  There
is also some variability in how the basin names are represented, so this data
set will need some work.
}
\keyword{datasets}
