% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcammaptools.R
\docType{data}
\name{map.basin235.simple}
\alias{map.basin235.simple}
\title{Simplified base map for 235 global water basins}
\format{Spatial data frame}
\usage{
map.basin235.simple
}
\description{
The same map as \code{map.basin235} but with only Polygons that have an area
greater than 2.5 square degrees and simplified Polygon borders.
}
\keyword{datasets}
