% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcammaptools.R
\docType{data}
\name{map.chn}
\alias{map.chn}
\title{Base map for 32-region GCAM with China subregions}
\format{Simple feature collection}
\usage{
map.chn
}
\description{
This map has the 32 GCAM regions, plus the subregions corresponding to
China's provinces, municipalities, autonomous regions, and SARs.
}
\keyword{datasets}
