% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcammaptools.R
\docType{data}
\name{map.chn.simple}
\alias{map.chn.simple}
\title{Simplified base map for 32-region GCAM with China subregions}
\format{Spatial data frame}
\usage{
map.chn.simple
}
\description{
The same map as \code{map.chn} but with only Polygons that have an area
greater than 2.5 square degrees and simplified Polygon borders.
}
\keyword{datasets}
