% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcammaptools.R
\docType{data}
\name{map.countries}
\alias{map.countries}
\title{Base map for gridded data over national borders}
\format{Simple feature collection}
\usage{
map.countries
}
\description{
This map has the administrative borders of the world's countries. It is meant
for use with gridded data only, as no GCAM output matches country boundaries.
}
\keyword{datasets}
