% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcammaptools.R
\docType{data}
\name{map.rgn14}
\alias{map.rgn14}
\title{Base map for 14-region GCAM}
\format{Spatial data frame}
\usage{
map.rgn14
}
\description{
This is the region map used in versions of GCAM prior to GCAM 4.0.  It is
largely obsolete, but there are still some data from those days floating
around in the wild
}
\keyword{datasets}

