% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcammaptools.R
\docType{data}
\name{map.rgn14.simple}
\alias{map.rgn14.simple}
\title{Simplified base map for 14-region GCAM}
\format{Simple feature collection}
\usage{
map.rgn14.simple
}
\description{
The same map as \code{map.rgn14} but with only Polygons that have an area
greater than 2.5 square degrees and simplified Polygon borders.
}
\keyword{datasets}
