% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcammaptools.R
\docType{data}
\name{map.rgn32}
\alias{map.rgn32}
\title{Base map for 32-region GCAM}
\format{Spatial data frame}
\usage{
map.rgn32
}
\description{
This is the region map used in GCAM 4.0 and subsequent.  This version of the
map does not include the Taiwan region.
}
\keyword{datasets}
