% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcammaptools.R
\docType{data}
\name{map.usa}
\alias{map.usa}
\title{Base map for 32-region GCAM with USA states}
\format{Simple feature collection}
\usage{
map.usa
}
\description{
This map has the 32 GCAM regions, plus the subregions corresponding to
US states including the District of Columbia.
}
\keyword{datasets}
