% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_header.R
\docType{data}
\name{na_aea}
\alias{na_aea}
\title{Proj4 string for the Albers equal area projection over North America.}
\format{An object of class \code{character} of length 1.}
\usage{
na_aea
}
\description{
String for specifying the Albers equal area projection over North
America in mapping functions.  This is a conic projection situatied
over the continental United States.  Its value is \code{'+proj=aea
+lat_1=20 +lat_2=60 +lat_0=40 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80
+datum=NAD83'}
}
\keyword{datasets}

