% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_header.R
\docType{data}
\name{ORIENTATION_LA}
\alias{ORIENTATION_LA}
\title{Orientation vector for orthographic projection of the Latin America superregion}
\format{An object of class \code{numeric} of length 3.}
\usage{
ORIENTATION_LA
}
\description{
This vector can be used as the \code{orientation} argument to
\code{\link{plot_GCAM}}.
}
\keyword{datasets}

