% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_header.R
\docType{data}
\name{ortho}
\alias{ortho}
\title{Projection string for orthographic projections}
\format{An object of class \code{character} of length 1.}
\usage{
ortho
}
\description{
Unlike most projections, the orthographic does not use proj4; it
uses the \code{mapproj::coord_map} function.  You can pass
this value to the \code{proj} argument of \code{\link{plot_GCAM}}
to get an orthographic projection.
}
\keyword{datasets}
