% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pgon_from_extent}
\alias{pgon_from_extent}
\title{Create a rectangluar sf polygon from numeric extent or sf bbox.}
\usage{
pgon_from_extent(ext)
}
\arguments{
\item{ext}{Numeric extent [xmin, xmax, ymin, ymax] or an object of class bbox}
}
\description{
Create a rectangluar sf polygon from numeric extent or sf bbox.
}
