% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{plot_GCAM_grid}
\alias{plot_GCAM_grid}
\title{Plot a gridded dataset over a base map}
\usage{
plot_GCAM_grid(plotdata, col, map = map.rgn32, proj = robin,
  proj_type = NULL, legend = F, alpha = 0.8, ...)
}
\arguments{
\item{plotdata}{Data frame with the coordinates and values to be plotted.
Must contain 'lat' and 'lon' columns.}

\item{col}{Name of the column holding the data values to plot}

\item{map}{Base map data.  Default is GCAM 32-region}

\item{proj}{Map projection to use in the display map.  This should be a proj4
string, except for a few special cases.  There are also symbols defined for
some frequently used projections (e.g. \code{\link{robin}} or
\code{\link{na_aea}}).}

\item{proj_type}{Either esri, epsg, or sr-org as string.  These correspond to
available reference types hosted by \url{http://spatialreference.org/}.}

\item{legend}{Boolean flag: True = display map legend; False = do not display
legend.}

\item{alpha}{Transparency of the grid data layer.  Given as a number between
0 and 1, where 0 is completely transparent and 1 is completely opaque.}

\item{...}{Other parameters passed on to \code{plot_GCAM}.}
}
\description{
This function produces a map visualization of a gridded (i.e., values
specified by latitude and longitude) data set.  The data will be plotted over
the base map supplied
}
\details{
The plot data should be in the form of a table of latitude (lat), longitude
(lon), and data values.  The name of the data column is given as an argument
to the function, so you can have, for example, latitude and longitude columns
followed by columns for time slices.  Columns besides the coordinate and data
columns will be ignored.
}
