% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{process_data}
\alias{process_data}
\title{Process data}
\usage{
process_data(map_data, data_key_field, data_col, bin_method, bins)
}
\arguments{
\item{map_data}{(Data Frame or Character) - A data frame that contains the output data to map, or alternatively a full path to a CSV}

\item{data_key_field}{(Character) - Name of key field in data_obj for merging with shape_data}

\item{data_col}{(Character) - Column name that contains the data object's output variable}

\item{bin_method}{(Character) - Method or function to use to split continuous data into discrete chunks (one of "quantile", "equal", "pretty", "kmeans") (default "pretty")}

\item{bins}{(Numeric) - Number of bins/segments in which to divide the raster}
}
\value{
(Data Frame or Character) - Returns the resulting simplified SF object or an error string if failed
}
\description{
Process data
}
