% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{process_shape}
\alias{process_shape}
\title{Process shape}
\usage{
process_shape(shape_data, simplify, shape_label_field)
}
\arguments{
\item{shape_data}{(SF, SP, or Character) - Either the full path string to a shape file (with included necessary files) or an SF shape object}

\item{simplify}{(Boolean) - Option to reduce the number/complexity of the polygons in the shape file (default FALSE)}

\item{shape_label_field}{(Character) - Optional field for plotting data available from the shape attributes/fields (such as country name)}
}
\value{
(SF or Character) - Returns the resulting simplified SF object or an error string if failed
}
\description{
Process shape
}
