% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{qualPalette}
\alias{qualPalette}
\title{Generate a color palette for categorical data.}
\usage{
qualPalette(n = 31, pal = "Set3", na.val = "grey50")
}
\arguments{
\item{n}{Number of entries desired for the palette}

\item{pal}{Name of the palette to base the new palette on. See
\code{RColorBrewer} for the palettes available.}

\item{na.val}{Color to use for the null region}
}
\description{
Generate a palette with a specified number of entries.  This
function uses a ramp function to extend the palettes from
\code{\link{RColorBrewer}} so they can handle a larger number of
entries.
}

