% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{remove_invalid}
\alias{remove_invalid}
\title{Clean spatial data.}
\usage{
remove_invalid(sfcobj)
}
\arguments{
\item{sfcobj}{Object of class \code{sf} or \code{sfc}}
}
\description{
Removes empty, corrupt, or invalid geometries from an sfc object.
}
