% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{reproject}
\alias{reproject}
\title{Transform the projection of an sf object}
\usage{
reproject(sdf, prj4s)
}
\arguments{
\item{sdf}{sf object}

\item{prj4s}{Proj4 string}
}
\description{
Recalculates the projection of the input sf object to
the user-defined one using a Proj4 string.
}
