% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\docType{data}
\name{rgn14}
\alias{rgn14}
\title{Designator for the rgn14 map set}
\format{An object of class \code{name} of length 1.}
\usage{
rgn14
}
\description{
This symbol will select the rgn14 map set
}
\keyword{datasets}

