% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\docType{data}
\name{rgn32}
\alias{rgn32}
\title{Designator for the rgn32 map set}
\format{An object of class \code{name} of length 1.}
\usage{
rgn32
}
\description{
This symbol will select the rgn32 map set
}
\keyword{datasets}

