% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_header.R
\docType{data}
\name{robin}
\alias{robin}
\title{Proj4 string for the Robinson World projection}
\format{An object of class \code{character} of length 1.}
\usage{
robin
}
\description{
String for specifying the Robinson projection in mapping functions.
Its value is \code{'+proj=robin'}
}
\keyword{datasets}
