% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{save_plot}
\alias{save_plot}
\title{Saves a plot to disk}
\usage{
save_plot(output_file = NULL, dpi = 150, map_width = 15, map_height = 10)
}
\arguments{
\item{output_file}{(Character) - Output file path and file name and type to save the resulting plot (e.g. "c:/temp/output.png") Available file types ("eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp", "svg")}

\item{dpi}{(Numeric) - Settable DPI for different print and screen formats (default 150)}

\item{map_width}{(Numeric) - Map width in inches (default 15)}

\item{map_height}{(Numeric) - Map height in inches (default 10)}
}
\value{
(Character) - Returns a character string with success or an error string if failed
}
\description{
Provides basic error handling and saving of the output map to file
}
\author{
Jason Evanoff, jason.evanoff@pnnl.gov
}
