% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{set_col}
\alias{set_col}
\title{Helper to set column name for NULL.}
\usage{
set_col(col, sub_val = 1)
}
\arguments{
\item{col}{Field name with target information to map.}

\item{sub_val}{Field position to substitue as column reference}
}
\description{
Set column name to substitute position when NULL.
}
