% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{set_color_scheme}
\alias{set_color_scheme}
\title{Set color scheme for ggplot object.}
\usage{
set_color_scheme(mapdata, col, qtitle, nacolor, colors, colorfcn)
}
\arguments{
\item{mapdata}{The sf object containing the spatial data.}

\item{col}{Field name with target information to map.}

\item{qtitle}{User specified title for color bar}

\item{nacolor}{Color for NA features}

\item{colors}{Color vector}

\item{colorfcn}{Function to create color scheme}
}
\description{
Build color scheme object for ggplot.
}
