% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spat_bb}
\alias{spat_bb}
\title{Create sf object from numeric extent.}
\usage{
spat_bb(b_ext, buff_dist = 0,
  proj4s = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs")
}
\arguments{
\item{b_ext}{Numeric extent [xmin, xmax, ymin, ymax]}

\item{buff_dist}{Distance in decimal degrees to expand the bounding box by in
all directions.}

\item{proj4s}{Either the proj4 string or EPSG number of the native projection
of the bounds}
}
\description{
Creates a sf object from numeric extent vector and applies a default WGS84
(EPSG:4326) coordinate reference system.
}
