% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{translateProvince}
\alias{translateProvince}
\title{Replace subregion abbreviations with full subregion names}
\usage{
translateProvince(datatable, provincefile)
}
\arguments{
\item{datatable}{The table with the abbreviated names in it.}

\item{provincefile}{Name of a defined mapset OR name of a file containing the
lookup table.}
}
\description{
Subregions are given two-letter abbreviations in GCAM output.  This function
uses a lookup table to restore the full names.
}
