% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\docType{data}
\name{usa}
\alias{usa}
\title{Designator for the usa map set}
\format{An object of class \code{name} of length 1.}
\usage{
usa
}
\description{
This symbol will select the usa map set
}
\keyword{datasets}
