% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_checking.R
\name{verify_csv}
\alias{verify_csv}
\title{Verify csv data}
\usage{
verify_csv(map_data)
}
\arguments{
\item{map_data}{(Data Frame or Character) - Either the full path string to data file (csv) or a data.frame object}
}
\value{
(Character) - Returns a success token or an error string if failed
}
\description{
This function runs a check on csv data inputs and looks for errors
}
