% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_checking.R
\name{verify_data}
\alias{verify_data}
\title{Verify map data}
\usage{
verify_data(map_data, data_key_field, data_col)
}
\arguments{
\item{map_data}{(Data Frame or Character) - A data frame that contains the output data to map, or alternatively a full path to a CSV}

\item{data_key_field}{(Character) - Name of key field in data_obj for merging with shape_data}

\item{data_col}{(Character) - Column name that contains the data object's output variable}
}
\value{
(Character) - Returns a success token or an error string if failed
}
\description{
This function runs a check on map data inputs and looks for errors
}
