% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_checking.R
\name{verify_map_params}
\alias{verify_map_params}
\title{Verify map parameters}
\usage{
verify_map_params(dpi, expand_xy, map_xy_min_max, map_title, map_palette,
  map_palette_reverse, map_palette_type, map_width_height_in,
  map_legend_title, map_x_label, map_y_label)
}
\arguments{
\item{dpi}{(Numeric) - Settable DPI for different print/screen formats (default 150)}

\item{expand_xy}{(c(Numeric, Numeric)) - Sets expansion amount for the X and Y scale of the map - Vector (expand x, expand y) (default c(0,0))}

\item{map_xy_min_max}{(c(Numeric, ...)) - Vector that describes the desired extent of the map in the form of (xmin, xmax, ymin, ymax) (default: c(-180, 180, -90, 90))}

\item{map_title}{(Character) - Title to be displayed on the output map}

\item{map_palette}{(Character) - Optional variable to manually set the colorscale to a specific palette from RColorbrewer}

\item{map_palette_reverse}{(Boolean) - Set palette to reverse direction TRUE/FALSE}

\item{map_width_height_in}{(c(Numeric, Numeric)) - Vector that describes the desired file size of the output image in the form of (width, height) in inches (default c(15, 10))}

\item{map_legend_title}{(Character) - Text for the legend header}

\item{map_x_label}{(Character) - Label for x axis (default Lon)}

\item{map_y_label}{(Character) - Label for y axis (default Lat)}
}
\description{
Verify map parameters
}
