% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_checking.R
\name{verify_raster}
\alias{verify_raster}
\title{Verify raster data}
\usage{
verify_raster(
  raster_data = NULL,
  raster_col = NULL,
  raster_band = 1,
  bin_method = "pretty",
  bins = 8,
  convert_zero = FALSE
)
}
\arguments{
\item{raster_data}{(Raster or Character) - Either the full path string to raster file or a raster object}

\item{raster_col}{(Character) - Column name that contains the raster object's output variable}

\item{raster_band}{(Numeric) - Future variable for dealing with multi band or time series rasters etc (default 1)}

\item{bin_method}{(Character) - Method or function to use to split continuous data into discrete chunks (default "pretty")}

\item{bins}{(Numeric) - Number of bins in which to divide the raster (default 8)}

\item{convert_zero}{(Boolean) - Convert raster zero values to NA (default FALSE)}
}
\value{
(Character) - Returns either "Success" or an error string if failed
}
\description{
This function runs a check on raster inputs and looks for errors, missing projection, or missing key fields
}
\author{
Jason Evanoff, jason.evanoff@pnnl.gov
}
