% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_checking.R
\name{verify_simplify_mapdata}
\alias{verify_simplify_mapdata}
\title{Verify simplify}
\usage{
verify_simplify_mapdata(
  map_data = NULL,
  min_area = 2.5,
  degree_tolerance = 0.1
)
}
\arguments{
\item{min_area}{(Numeric) Minimum area of polygons to keep.}

\item{degree_tolerance}{(Numeric) Tolerance parameter for simplifying polygons.}

\item{mapdata}{(sf) object containing polygons or multipolygons to simplify.}
}
\value{
(Character) - Returns a success token or an error string if failed
}
\description{
This function does simple verification for the simplify_mapdata function
}
