% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_header.R
\docType{data}
\name{wgs84}
\alias{wgs84}
\title{Proj4 string for default WGS84 (EPSG:4326) coordinate reference system}
\format{An object of class \code{character} of length 1.}
\usage{
wgs84
}
\description{
String for specifying the default WGS84 projection in mapping functions
Its value is \code{'+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs'}
}
\keyword{datasets}
