% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_header.R
\docType{data}
\name{wintri}
\alias{wintri}
\title{Proj4 string for the Winkel-Tripel World projection}
\format{An object of class \code{character} of length 1.}
\usage{
wintri
}
\description{
String for specifying the Winkel-Tripel projection in mapping
functions.  Its value is \code{'+proj=wintri'}
}
\keyword{datasets}

