% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{zoom_bounds}
\alias{zoom_bounds}
\title{Build the coordinate zooming bounds for ggplot object.}
\usage{
zoom_bounds(mapdata, bbox, extent, p4s)
}
\arguments{
\item{mapdata}{The sf object containing the spatial data.}

\item{bbox}{Bounding box.}

\item{extent}{Extent provided by the user or as default.}

\item{p4s}{Proj4 string set by user}
}
\description{
Either take the bounding box or the map coordinates depending on extent.
}
