% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{choropleth}
\alias{choropleth}
\title{Create a choropleth map object from shape and data object and return, save (optional) the output}
\usage{
choropleth(shape_data = NULL, shape_key_field = NULL,
  shape_label_field = NULL, shape_label_size_field = "1",
  shape_data_field = NULL, shape_xy_fields = c("LON", "LAT"),
  shape_geom_field = "geometry", simplify = FALSE, map_data = NULL,
  data_key_field = NULL, data_col = NULL, bin_method = "pretty",
  bins = NULL, dpi = 150, output_file = NULL, expand_xy = c(0, 0),
  map_xy_min_max = c(-180, 180, -90, 90), map_title = NULL,
  map_palette = NULL, map_palette_reverse = FALSE,
  map_palette_type = "seq", map_width_height_in = c(15, 10),
  map_legend_title = NULL, map_x_label = "Lon", map_y_label = "Lat")
}
\arguments{
\item{shape_data}{(SF, SP, or Character) - Either the full path string to a shape file (with included necessary files) or an SF shape object}

\item{shape_key_field}{(Character) - Name of key field in shape object for merging with map_data object}

\item{shape_label_field}{(Character) - Optional field for plotting data available from the shape attributes/fields (such as country name)}

\item{shape_label_size_field}{(Character) - Optional field used for computing shape label size dynamically (ie by area or amount etc.)}

\item{shape_data_field}{(Character) - Optional field for utilizing a field within the shape data as the map data field. Negates the map_data variable}

\item{shape_xy_fields}{(c(Character, Character)) - Vector that specifies the x and y field names in the shape object (default c("LAT", "LON"))}

\item{shape_geom_field}{(Character) - Specifies field within shape object that contains needed geometry (default "geometry")}

\item{simplify}{(Boolean) - Option to reduce the number/complexity of the polygons in the shape file (default FALSE)}

\item{map_data}{(Data Frame or Character) - A data frame that contains the output data to map, or alternatively a full path to a CSV}

\item{data_key_field}{(Character) - Name of key field in data_obj for merging with shape_data}

\item{data_col}{(Character) - Column name that contains the data object's output variable}

\item{bin_method}{(Character) - Method or function to use to split continuous data into discrete chunks (one of "quantile", "equal", "pretty", "kmeans") (default "pretty")}

\item{bins}{(Numeric) - Number of bins/segments in which to divide the raster}

\item{dpi}{(Numeric) - Settable DPI for different print/screen formats (default 150)}

\item{output_file}{(Character) - Output file path and file name/type to save the resulting plot (e.g. "c:/temp/output.png") (Types accepted: "pdf", "jpeg", "tiff", "png", "bmp", "svg", "eps", "ps", "tex") (default PNG)}

\item{expand_xy}{(c(Numeric, Numeric)) - Sets expansion amount for the X and Y scale of the map - Vector (expand x, expand y) (default c(0,0))}

\item{map_xy_min_max}{(c(Numeric, ...)) - Vector that describes the desired extent of the map in the form of (xmin, xmax, ymin, ymax) (default: c(-180, 180, -90, 90))}

\item{map_title}{(Character) - Title to be displayed on the output map}

\item{map_palette}{(Character) - Optional variable to manually set the colorscale to a specific palette from RColorbrewer}

\item{map_palette_reverse}{(Boolean) - Set palette to reverse direction TRUE/FALSE}

\item{map_palette_type}{(Character) - Variable to load default palette by type of data ("qual" for qualitative data, "seq" for sequential data, "div" for divergent data) (default "seq")}

\item{map_width_height_in}{(c(Numeric, Numeric)) - Vector that describes the desired file size of the output image in the form of (width, height) in inches (default c(15, 10))}

\item{map_legend_title}{(Character) - Text for the legend header}

\item{map_x_label}{(Character) - Label for x axis (default Lon)}

\item{map_y_label}{(Character) - Label for y axis (default Lat)}
}
\value{
(ggplot2 or Character) - Returns a ggplot object of the resulting map or an error string if failed
}
\description{
Choropleth map
}
