% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{create_map}
\alias{create_map}
\title{Create a map object and return/save the output}
\usage{
create_map(shape_data = NULL, raster_data = NULL, raster_col = NULL,
  simplify = FALSE, raster_band = 1, dpi = 150, output_file = NULL,
  output_file_type = "png", expand_xy = c(0, 0),
  map_xy_min_max = c(-180, 180, -90, 90), map_title = NULL,
  map_palette = "RdYlBu", map_width_height_in = c(15, 10),
  map_legend_title = NULL, map_x_label = "x", map_y_label = "y")
}
\arguments{
\item{shape_data}{Either the full path string to shape file or an SF shape object}

\item{raster_data}{Either the full path string to raster file or a raster object}

\item{raster_col}{Column that contains the raster object's output variable}

\item{simplify}{Option to reduce the number/complexity of the polygons in the shape file}

\item{raster_band}{Future variable for dealing with multi band/time series rasters}

\item{dpi}{Settable DPI for different print/screen formats}

\item{output_file}{Output file path to save the resulting plot}

\item{output_file_type}{Sets default file type for saving plots. Choices restricted to those in ggsave function}

\item{expand_xy}{Sets expansion amount for the X and Y scale of the map - Vector (expand x, expand y)}

\item{map_xy_min_max}{Vector that describes the desired extent of the map in the form of (xmin, xmax, ymin, ymax)}

\item{map_title}{Title to be displayed on the output map}

\item{map_palette}{Variable to hold the type of colorscale to be used}

\item{map_width_height_in}{Vector that describes the desired file size of output map image in the form of (width, height) in inches}

\item{map_legend_title}{Text for the legend header}

\item{map_x_label}{Label for x axis}

\item{map_y_label}{Label for y axis}
}
\value{
A ggplot object of the resulting map
}
\description{
This function is designed to take both a shape and raster object or path and create a standardized, congruent, output map.
}
