% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{create_map}
\alias{create_map}
\title{Create a map object and return/save the output}
\usage{
create_map(shape_data = NULL, shape_label_field = NULL,
  simplify = FALSE, raster_data = NULL, raster_col = NULL,
  raster_band = 1, bin_method = NULL, bins = NULL,
  convert_zero = FALSE, dpi = 150, output_file = NULL,
  expand_xy = c(0, 0), map_xy_min_max = c(-180, 180, -90, 90),
  map_title = NULL, map_palette = "RdYlBu",
  map_width_height_in = c(15, 10), map_legend_title = NULL,
  map_x_label = "Lon", map_y_label = "Lat")
}
\arguments{
\item{shape_data}{(SF, SP, or Character) - Either the full path string to shape file or an SF shape object}

\item{shape_label_field}{(Character) - Optional field for plotting data available from the shape attributes/fields}

\item{simplify}{(Boolean) - Option to reduce the number/complexity of the polygons in the shape file}

\item{raster_data}{(Raster or Character) - Either the full path string to raster file or a raster object}

\item{raster_col}{(Character) - Column name that contains the raster object's output variable}

\item{raster_band}{(Numeric) - Future variable for dealing with multi band/time series rasters etc}

\item{bin_method}{(Character) - Method or function to use to split continuous data into discrete chunks}

\item{bins}{(Numeric) - Number of bins/segments in which to divide the raster}

\item{convert_zero}{(Boolean) - Convert raster 0 values to NA, default FALSE}

\item{dpi}{(Numeric) - Settable DPI for different print/screen formats}

\item{output_file}{(Character) - Output file path and file name/type to save the resulting plot (e.g. "c:/temp/output.png") Available file types ("eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp", "svg")}

\item{expand_xy}{(c(Numeric, Numeric)) - Sets expansion amount for the X and Y scale of the map - Vector (expand x, expand y)}

\item{map_xy_min_max}{(c(Numeric, Numeric, ...)) - Vector that describes the desired extent of the map in the form of (xmin, xmax, ymin, ymax)}

\item{map_title}{(Character) - Title to be displayed on the output map}

\item{map_palette}{(Character) - Variable to hold the type of colorscale to be used}

\item{map_width_height_in}{(c(Numeric, Numeric)) - Vector that describes the desired file size of output map image in the form of (width, height) in inches}

\item{map_legend_title}{(Character) - Text for the legend header}

\item{map_x_label}{(Character) - Label for x axis}

\item{map_y_label}{(Character) - Label for y axis}
}
\value{
(ggplot2 or Character) - Returns a ggplot object of the resulting map or an error string if failed
}
\description{
This function is designed to take both a shape and raster object or path and create a standardized, congruent, output map.
}
