% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{filter_spatial}
\alias{filter_spatial}
\title{Get features topologically associated with extent bounds.}
\usage{
filter_spatial(mapdata, bbox, agr_type = "constant",
  topo = sf::st_intersects)
}
\arguments{
\item{mapdata}{The sf object containing the spatial data.}

\item{bbox}{Bounding box.}

\item{agr_type}{Inherited attribute-geometry-relationship type from plot_GCAM
function params.}

\item{topo}{SF topologic function to define how the join will be conducted.
Default is to join any feature that intersects the bounding box.}
}
\description{
Conducts a spatial join to retrieve spatial features that are topologically associated
(intersects, contains, within, etc.) with the provided bounds.
}
