% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{get_prj4s}
\alias{get_prj4s}
\title{Retrieve proj4 projection string.}
\usage{
get_prj4s(obj = NULL, prj4s_key = NULL, epsg = NULL, esri = NULL,
  srorg = NULL, lu = NULL)
}
\arguments{
\item{obj}{Use object instead that has a predefined proj4 string.}

\item{prj4s_key}{Lookup key string identifying the default projection.}

\item{epsg}{The EPSG projection code as an integer.}

\item{esri}{The ESRI projection code as an integer.}

\item{srorg}{The SR-ORG projection code as an integer.}

\item{lu}{A key = value list where key is a string and value is the
associated proj4 string.}
}
\description{
Provides a lookup list for default proj4 strings utilized.  Users
may also specify their own lookup list.  Options also include
providing either the EPSG, ESRI, or SR-ORG projection codes to
retrieve the associated proj4 string from a web query from
http://spatialreference.org.  Definitions for proj4 string
parmeters can be referenced here:
http://proj4.org/parameters.html#parameter-list
}
