% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{import_mapdata}
\alias{import_mapdata}
\title{Single import function for compatible data types.}
\usage{
import_mapdata(obj, fld = NULL, prj4s = wgs84)
}
\arguments{
\item{obj}{Input full path string or object.}

\item{fld}{Field name to use as identifier.}

\item{prj4s}{Proj4 string for projection (default WGS84).}
}
\value{
An sf object representation of the map data.
}
\description{
Imports available for sf objects, spatial data frames, ESRI Shapefiles, or
GeoJSON files.
}
