% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{load_shp}
\alias{load_shp}
\title{Import ESRI Shapefile or GeoJSON as sf object.}
\usage{
load_shp(file_pth)
}
\arguments{
\item{file_pth}{Full path to shapefile with extention (.shp).  Shapefiles must contain at least
.shp, .shx, and .dbf file to function properly.  For the purposes of this package, a .prj
file is also required to identify the native projection of the shapefile.}
}
\description{
Creates a Simple Feature (sf) object from full path string to ESRI Shapefile or
GeoJSON file. User defines which field is supposed to represent the ID for the data.
}
