% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{load_txt}
\alias{load_txt}
\title{Import text file as a data frame .}
\usage{
load_txt(txt, field)
}
\arguments{
\item{txt}{Full path to file including extension}
}
\description{
Creates a DataFrame from full path string to file.
}
