% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{process_data}
\alias{process_data}
\title{Process data}
\usage{
process_data(map_data)
}
\arguments{
\item{map_data}{(Data Frame or Character) - Either the full path string to data file (csv) or a data.frame object}
}
\value{
(Data Frame or Character) - Returns the resulting simplified SF object or an error string if failed
}
\description{
Process data
}
