% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{process_raster}
\alias{process_raster}
\title{Process raster}
\usage{
process_raster(raster_data, raster_col, raster_band, bin_method, bins,
  convert_zero)
}
\arguments{
\item{raster_data}{(Raster or Character) - Either the full path string to raster file or a raster object}

\item{raster_col}{(Character) - Column name that contains the raster object's output variable}

\item{raster_band}{(Numeric) - Future variable for dealing with multi band/time series rasters etc}

\item{bin_method}{(Character) - Method or function to use to split continuous data into discrete chunks}

\item{bins}{(Numeric) - Number of bins/segments in which to divide the raster}

\item{convert_zero}{(Boolean) - Convert raster zero values to NA}
}
\value{
(Raster or Character) - Returns the resulting raster object or an error string if failed
}
\description{
Process raster
}
