% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_checking.R
\name{return_error}
\alias{return_error}
\title{Return and output errors}
\usage{
return_error(error, location)
}
\arguments{
\item{error}{(Character) - Specific error string for output}

\item{location}{(Character) - Location the error was caught}
}
\value{
(Character) - Prints the error to console and also returns it to caller
}
\description{
This function prints out any caught errors to the console and returns the custom error string to the calling function
}
\author{
Jason Evanoff, jason.evanoff@pnnl.gov
}
