% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{save_plot}
\alias{save_plot}
\title{Handles saving the map output to disk}
\usage{
save_plot(output_file, dpi, map_width, map_height)
}
\arguments{
\item{output_file}{(Character) - Output file path and file name/type to save the resulting plot (e.g. "c:/temp/output.png") Available file types ("eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp", "svg")}

\item{dpi}{(Numeric) - Settable DPI for different print/screen formats}

\item{map_width}{(Numeric) - Map width in inches}

\item{map_height}{(Numeric) - Map height in inches}
}
\value{
(Character) - Returns a character string with success or an error string if failed
}
\description{
Initially one line of code, this function is likely to be expanded
}
