% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_Functions.R
\name{set_col}
\alias{set_col}
\title{Helper to set column name for NULL.}
\usage{
set_col(col)
}
\arguments{
\item{col}{Field name with target information to map.}

\item{sub_val}{Field position to substitue as column reference}
}
\value{
The name of the data column or 1 if NULL.
}
\description{
Set column name to substitute position when NULL.
}
