% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplify_mapdata}
\alias{simplify_mapdata}
\title{Reduce number of polygons and size of polygons for map Shapefiles}
\usage{
simplify_mapdata(mapdata, min_area = 2.5, degree_tolerance = 0.1)
}
\arguments{
\item{mapdata}{sf object containing polygons or multipolygons to simplify.}

\item{min_area}{Minimum area of polygons to keep.}

\item{degree_tolerance}{Tolerance parameter for simplifying polygons.}
}
\value{
The simplified sf object.
}
\description{
Takes a sf object representation of a map and simplifys it by removing
polygons that are under a certain size.
}
\details{
** NOTE: This function adds two polygons to the edges of the map to prevent
         the removal of polygons near the edges redefining the map bounds.
}
