% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_checking.R
\name{verify_shape}
\alias{verify_shape}
\title{Error checking functions for maps package
Verify shape data}
\usage{
verify_shape(
  shape_data,
  simplify = FALSE,
  shape_label_field = NULL,
  shape_data_field = NULL,
  shape_key_field = NULL,
  shape_label_size = "1",
  shape_xy_fields = c("LAT", "LON"),
  shape_geom_field = "geometry"
)
}
\arguments{
\item{shape_data}{(SF, SP, or Character) - Either the full path string to a shape file (with included necessary files) or an SF shape object}

\item{simplify}{(Boolean) - Option to reduce the number or complexity of the polygons in the shape file (default FALSE)}

\item{shape_label_field}{(Character) - Optional field for plotting data available from the shape attributes & fields (such as country name)}

\item{shape_data_field}{(Character) - Optional field for utilizing a field within the shape data as the map data field. Negates the map_data variable}

\item{shape_key_field}{(Character) - Name of key field in shape object for merging with map_data object}

\item{shape_xy_fields}{(c(Character, Character)) - Vector that specifies the x and y field names in the shape object (default c("LAT", "LON"))}

\item{shape_geom_field}{(Character) - Specifies field within shape object that contains needed geometry (default "geometry")}

\item{shape_label_size_field}{(Character) - Optional field used for computing shape label size dynamically (ie by area or amount etc.)}
}
\value{
(Character) - Returns either "Success" or an error string if failed
}
\description{
This function runs a check on shape inputs and looks for errors
}
\author{
Jason Evanoff, jason.evanoff@pnnl.gov
}
